+++ core/src/linuxX64Main/kotlin/LedComponents.kt
@@ -6,6 +6,7 @@ interface LedComponent {
     var color: Color
     var speed: Int
     var brightness: Int
+    var useRandomColor: Boolean
 
     fun assignValuesFromChannel(channelValues: ChannelValues)
 }
@@ -21,6 +22,9 @@ interface BasicLedComponent : LedComponent {
         if (mode != LedMode.CYCLE) {
             color = channelValues.color
         }
+        if (mode.colorSupport == ColorSupport.ALL) {
+            useRandomColor = channelValues.colorSource == 0x80u.toUByte()
+        }
         speed = mode.speeds.indexOfOrNull(channelValues.speed)?.plus(1) ?: 3
         brightness = mode.brightnesses.indexOfOrNull(channelValues.brightness)?.plus(1) ?: 2
     }
@@ -30,7 +34,8 @@ interface BasicLedComponent : LedComponent {
         get() {
             val brightness = mode.brightnesses.elementAtOrNull(brightness - 1) ?: 0u
             val speed = mode.speeds.elementAtOrNull(speed - 1) ?: 0x2Cu
-            return ubyteArrayOf(channel, speed, 0x20u, mode.mode, 0xFFu, brightness, *color.bytes)
+            val colorSource: UByte = if (useRandomColor) 0x80u else 0x20u
+            return ubyteArrayOf(channel, speed, colorSource, mode.mode, 0xFFu, brightness, *color.bytes)
         }
 }
 
@@ -39,6 +44,7 @@ class LogoComponent(initialValues: ChannelValues) : BasicLedComponent {
     override val channel: UByte = 0x05u
     override lateinit var mode: LedMode
     override var color = Color(0, 0, 0)
+    override var useRandomColor = false
     override var speed = 0
     override var brightness = 0
 
@@ -52,6 +58,7 @@ class FanComponent(initialValues: ChannelValues) : BasicLedComponent {
     override val channel: UByte = 0x06u
     override lateinit var mode: LedMode
     override var color = Color(0, 0, 0)
+    override var useRandomColor = false
     override var speed = 0
     override var brightness = 0
     var mirage = false
@@ -68,6 +75,7 @@ enum class RotationDirection(val value: UByte) { CLOCKWISE(0u), COUNTERCLOCKWISE
 class RingComponent(initialValues: ChannelValues) : LedComponent {
     lateinit var mode: RingMode
     override lateinit var color: Color
+    override var useRandomColor = false
     override var speed = 0
     override var brightness = 0
     lateinit var direction: RotationDirection
@@ -77,14 +85,17 @@ class RingComponent(initialValues: ChannelValues) : LedComponent {
     }
 
     override fun assignValuesFromChannel(channelValues: ChannelValues) {
-        mode = RingMode.values().first { it.channel == channelValues.channel && it.mode == channelValues.mode }
-        color = channelValues.let { if (mode.supportsColor) it.color else Color(0u, 0u, 0u) }
+        mode = RingMode.values().first { it.channel == channelValues.channel }
+        color = channelValues.let { if (mode.colorSupport != ColorSupport.NONE) it.color else Color(0u, 0u, 0u) }
         speed = mode.speeds.indexOfOrNull(channelValues.speed)?.plus(1) ?: 3
         brightness = mode.brightnesses.indexOfOrNull(channelValues.brightness)?.plus(1) ?: 2
         direction = if (mode.supportsDirection) {
             RotationDirection.values()[channelValues.colorSource.toInt() and 1]
         } else {
             RotationDirection.CLOCKWISE
+        }
+        if (mode.colorSupport == ColorSupport.ALL) {
+            useRandomColor = channelValues.colorSource and 0x80u == 0x80u.toUByte()
         }
     }
 
@@ -94,7 +105,12 @@ class RingComponent(initialValues: ChannelValues) : LedComponent {
             val speed = if (mode != RingMode.MORSE) {
                 mode.speeds.elementAtOrNull(speed - 1) ?: 0xFFu
             } else 0x6Bu
-            val colorSource = if (mode.supportsDirection) direction.value else mode.colorSource
+            val colorSource = when {
+                mode.colorSupport == ColorSupport.ALL && useRandomColor ->
+                    0x80u.let { if (mode.supportsDirection) it + direction.value else it }.toUByte()
+                mode.supportsDirection -> direction.value
+                else -> mode.colorSource
+            }
             return ubyteArrayOf(mode.channel, speed, colorSource, mode.mode, 0xFFu, brightness, *color.bytes)
         }
 }
@@ -102,17 +118,19 @@ class RingComponent(initialValues: ChannelValues) : LedComponent {
 @OptIn(ExperimentalUnsignedTypes::class)
 private fun List<UByte>.indexOfOrNull(value: UByte) = indexOf(value).let { if (it == -1) null else it }
 
+enum class ColorSupport { NONE, SPECIFIC, ALL }
+
 @OptIn(ExperimentalUnsignedTypes::class)
 enum class LedMode(
     val mode: UByte,
     val brightnesses: List<UByte> = listOf(0x4Cu, 0x99u, 0xFFu),
     val speeds: List<UByte> = emptyList(),
-    val supportsColor: Boolean = false
+    val colorSupport: ColorSupport = ColorSupport.NONE
 ) {
     OFF(0x00u, emptyList()),
-    STATIC(0x01u, supportsColor = true),
+    STATIC(0x01u, colorSupport = ColorSupport.SPECIFIC),
     CYCLE(0x02u, listOf(0x10u, 0x40u, 0x7Fu), listOf(0x96u, 0x8Cu, 0x80u, 0x6Eu, 0x68u)),
-    BREATHE(0x03u, speeds = listOf(0x3Cu, 0x37u, 0x31u, 0x2Cu, 0x26u), supportsColor = true)
+    BREATHE(0x03u, speeds = listOf(0x3Cu, 0x37u, 0x31u, 0x2Cu, 0x26u), colorSupport = ColorSupport.ALL)
 }
 
 @OptIn(ExperimentalUnsignedTypes::class)
@@ -120,25 +138,25 @@ enum class RingMode(
     val channel: UByte, val mode: UByte,
     val brightnesses: List<UByte> = listOf(0x4Cu, 0x99u, 0xFFu),
     val speeds: List<UByte> = emptyList(),
-    val supportsColor: Boolean = false,
+    val colorSupport: ColorSupport = ColorSupport.NONE,
     val supportsDirection: Boolean = false,
     val colorSource: UByte = 0x20u
 ) {
     OFF(0xFEu, 0x00u, emptyList(), emptyList()),
-    STATIC(0x00u, 0xFFu, supportsColor = true),
+    STATIC(0x00u, 0xFFu, colorSupport = ColorSupport.SPECIFIC),
     RAINBOW(0x07u, 0x05u, speeds = listOf(0x72u, 0x68u, 0x64u, 0x62u, 0x61u), colorSource = 0u),
     SWIRL(
         0x0Au, 0x4Au, speeds = listOf(0x77u, 0x74u, 0x6Eu, 0x6Bu, 0x67u),
-        supportsColor = true, supportsDirection = true
+        colorSupport = ColorSupport.ALL, supportsDirection = true
     ),
     CHASE(
         0x09u, 0xC3u, speeds = listOf(0x77u, 0x74u, 0x6Eu, 0x6Bu, 0x67u),
-        supportsColor = true, supportsDirection = true
+        colorSupport = ColorSupport.ALL, supportsDirection = true
     ),
     BOUNCE(0x08u, 0xFFu, speeds = listOf(0x77u, 0x74u, 0x6Eu, 0x6Bu, 0x67u), colorSource = 0x80u),
-    MORSE(0x0Bu, 0x05u, supportsColor = true, colorSource = 0u),
+    MORSE(0x0Bu, 0x05u, colorSupport = ColorSupport.ALL, colorSource = 0u),
     CYCLE(0x02u, 0xFFu, listOf(0x10u, 0x40u, 0x7Fu), listOf(0x96u, 0x8Cu, 0x80u, 0x6Eu, 0x68u)),
-    BREATHE(0x01u, 0xFFu, speeds = listOf(0x3Cu, 0x37u, 0x31u, 0x2Cu, 0x26u), supportsColor = true)
+    BREATHE(0x01u, 0xFFu, speeds = listOf(0x3Cu, 0x37u, 0x31u, 0x2Cu, 0x26u), colorSupport = ColorSupport.ALL)
 }
 
 class Color(val r: Int, val g: Int, val b: Int) {